chore: move repl integration tests into separate file
